Add snapcraft.yaml defining the snap package and its components
The DUB package uses `classic` confinement to ensure that it can access
the resources of the host system it is installed on, e.g. any installed
D compilers, libraries required for linking, and so on.

A simple `prepare:` scriptlet is used to invoke DUB's `build.sh` using
`dmd` as the compiler.  The `dump` plugin is then used (together with a
simple `stage:` rule) to include the compiled `bin/dub` in the package.

A fresh copy of the latest stable DMD is built to provide the D compiler
with which dub is built.  This is a bit of duplication of effort, but in
the short term should provide the quickest way to get a working compiler
for the appropriate target architecture.  A custom `dmd.conf` has been
included in order to ensure that the compiler finds druntime and phobos.

Note that DMD is not included in the final package, but merely used as a
bootstrapping mechanism to build DUB.