Update DUB to v1.5.0 and DMD to v2.076.0
This patch brings the packaged DUB and the DMD used to build it up to
date with their latest stable releases.

The v2.076.0 release expects druntime and phobos' make commands to be
provided with a `DMD_DIR` variable instead of just a path to the `dmd`
command, so `make-parameters` have been tweaked accordingly.

Fixes https://github.com/dlang-snaps/dub.snap/issues/4.