Bundle libcurl with the snap package
DUB uses `std.net.curl` to download package data, which in turn relies
on being able to dynamically load libcurl at runtime via `dlopen`.  If
libcurl is not available in the snap, but only on the host system, then
DUB may not be able to locate it: this happens e.g. on Ubuntu 18.04 and
quite possibly on other distros too.

This patch adds a separate `curl` part that includes `libcurl3-gnutls`
via `stage-packages`.  This is simpler than adding it to the `dub` part
as the `dump` plugin will not automatically stage packages specified in
this way.

Besides the libraries themselves, we need to add extra `LD_LIBRARY_PATH`
settings to the app environment to enable dub to find them.

The `license.txt` file has been updated to note the inclusion of these
libraries and point the reader to copyright and licensing information.

Adapted from an earlier PR by @philburr:
https://github.com/dlang-snaps/dub.snap/pull/13