Add base and license specifiers to snapcraft.yaml
See https://snapcraft.io/docs/base-snaps for more details on base snaps.
The `core` base is chosen to ensure compatibility with Ubuntu 16.04.

The license spec is chosen to reflect the license of DUB itself and the
druntime/phobos code required to build it.  System libraries are taken
as given, as their license details are provided in `usr/share/doc`.

This change has a few notable impacts:

  * we need to replace the old `prepare:` scriptlet with the new
    `override-build` standard, and invoke `snapcraftctl build` to
    ensure that the install stage completes successfully

  * we can no longer build with snapcraft versions earlier than 3
    or use the `snapcraft cleanbuild` command

  * by default snapcraft will now attempt to build using `multipass`
    (`snapcraft --use-lxd` will reproduce the old `cleanbuild`-style
    behaviour, while `snapcraft --destructive-mode` will build in the
    local environment)