Add a separate part to install the LICENSE file
The old habit of including a `license.txt` in the `snap` directory has
been deprecated for years (to the extent that it was ever even tacitly
approved of), and risks the license file not being included in the snap
package itself.

Adding an explicit `dub-license` part ensures that the license will
always be included.